+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.2.0]
+- Added an option to retrieve closed tickets, including the latest tag at the time the issue was closed.
+- Added an option to only retrieve tickets after a certain date.
+
 ## [1.1.0]
 - Include assignees on each issue in the CSV.